pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "revolt_mpd_status"
version = "0.0.1"
authors = [
    { name = "Amy Pérez", email = "amy+heya@amogus.cloud" }
]
dependencies = [
    "revolt.py ~= 0.2.0",
    "python-mpd2 ~= 3.1.1",
    "docopt-ng ~= 0.9.0"
]
description = "Show friends what you are listening on MPD"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]

[project.scripts]
revoltmpd = "revolt_mpd_status:main"

[project.urls]
Homepage = "https://github.com/amycatgirl/mpdrevoltstatus"
Issues = "https://github.com/amycatgirl/mpdrevoltstatus/issues"